Reference

​

Contents

​

Reference

Contents

Index

​

Index

​

NLPModels.grad!

NLPModels.obj

PartitionedKnetNLPModels.PS_deduction

PartitionedKnetNLPModels.precompile_ps_struct

PartitionedKnetNLPModels.ps

PartitionedKnetNLPModels.ps

​

NLPModels.grad! — Method

g = grad!(nlp, x, g)

Evaluate ∇f(x), the gradient of the objective function at x in place.

NLPModels.obj — Method

f = obj(nlp, x)

Evaluate `f(x)`, the objective function of `nlp` at `x`.

PartitionedKnetNLPModels.PS_deduction — Method

PS_deduction(c,dp)

Compute the partially separable structure of a network represented by the Chain c by performing a forward evaluation.

PartitionedKnetNLPModels.precompile_ps_struct — Method

precompile_ps_struct(network<:Chain)

The function is called on a network defined by Dense/Seplayer/Conv or an other layer that define also psstruct(layer); index. The result is a precompiled PS structure. Il faut déterminer pour chaque couche une fonction ps_struct qui déterminer l'information nécessaire à prendre en compte pour déterminer la séparabilité partielle du réseau. Sa forme est très proche de celle du réseau initial, mais chaque arc (ie variable) contient l'indice de la variable. Attention il ne faut pas oublier de stocker les biais qui sont également des variables. L'objectif est d'ensuite appelé une fonction évaluant rapidement la structure PS du réseau., par propagation des dépendances sur une évaluation backward. Pour une analyse de la structure sous influence du dropout j'ai intégré des vecteur booléen indiquant neurones de la couche précédente étaient actifs ou non. Cela se modélise par le vecteur dp dans PsDense

PartitionedKnetNLPModels.ps — Method

ps(psd::PsDense, Dep::Vector{Vector{int}})

pour chaque noeud on va ajouter les dépendances du layer propre (excepté dropout) on y ajoute ensuite les dépendances (car dense) de la couche précédente modélisé par Di (à modérer avec le dropout) Le cas de la couche Dense est particulier, sans dropout il n'a que peu d'intérêt

PartitionedKnetNLPModels.ps — Method

ps(pss::PsSep, Dep::Vector{Vector{int}})

pour chaque noeud on va ajouter les dépendances du layer propre: une sous partie de la couche précédente (+ dropout) on y ajoute ensuite certaines dépendances de la couche précédente modélisé par Di (à modérer avec le dropout) Dans le cas PS, la propagation des variables sera moins intense que le cas Dense